Externalize surface/device/etc
