+++ examples/gltf/src/main.rs
@@ -62,15 +62,24 @@ fn main() {
 
     let window_size = window.inner_size();
 
-    let renderer = pollster::block_on(
-        rend3::RendererBuilder::new(rend3::InternalSurfaceOptions {
-            vsync: rend3::VSyncMode::On,
-            size: UVec2::new(window_size.width, window_size.height),
-        })
-        .window(&window)
-        .build(),
-    )
-    .unwrap();
+    // Create the Instance, Adapter, and Device. We can specify preferred backend, device name, or rendering mode. In this case we let rend3 choose for us.
+    let iad = pollster::block_on(rend3::create_iad(None, None, None)).unwrap();
+
+    // The one line of unsafe needed. We just need to guarentee that the window outlives the use of the surface.
+    let surface = unsafe { iad.instance.create_surface(&window) };
+    // Get the preferred format for the surface.
+    let format = surface.get_preferred_format(&iad.adapter).unwrap();
+    // Configure the surface to be ready for rendering.
+    rend3::configure_surface(
+        &surface,
+        &iad.device,
+        format,
+        UVec2::new(window_size.width, window_size.height),
+        rend3::types::PresentMode::Mailbox,
+    );
+
+    // Make us a renderer.
+    let renderer = rend3::Renderer::new(iad).unwrap();
 
     // Create the pbr pipeline with the same internal resolution and 4x multisampling
     let mut routine = rend3_pbr::PbrRenderRoutine::new(
@@ -79,6 +88,7 @@ fn main() {
             resolution: UVec2::new(window_size.width, window_size.height),
             samples: rend3_pbr::SampleCount::Four,
         },
+        format,
     );
 
     // Create mesh and calculate smooth normals based on vertices.
@@ -130,22 +140,31 @@ fn main() {
             ..
         } => {
             let size = UVec2::new(size.width, size.height);
-            renderer.set_internal_surface_options(rend3::InternalSurfaceOptions {
-                vsync: rend3::VSyncMode::Off,
-                size,
-            });
+            // Reconfigure the surface for the new size.
+            rend3::configure_surface(
+                &surface,
+                &renderer.device,
+                format,
+                UVec2::new(size.x, size.y),
+                rend3::types::PresentMode::Mailbox,
+            );
+            // Tell the renderer about the new aspect ratio.
+            renderer.set_aspect_ratio(size.x as f32 / size.y as f32);
+            // Resize the internal buffers to the same size as the screen.
             routine.resize(
                 &renderer,
                 rend3_pbr::RenderTextureOptions {
                     resolution: size,
-                    samples: rend3_pbr::SampleCount::Four,
+                    samples: rend3_pbr::SampleCount::One,
                 },
             );
         }
         // Render!
         winit::event::Event::MainEventsCleared => {
+            // Get a frame
+            let frame = rend3::util::output::OutputFrame::from_surface(&surface).unwrap();
             // Dispatch a render!
-            let _stats = renderer.render(&mut routine, rend3::util::output::RendererOutput::InternalSurface);
+            let _stats = renderer.render(&mut routine, frame);
         }
         // Other events we don't care about
         _ => {}
