+++ rend3/Cargo.toml
@@ -47,7 +47,6 @@ log = "0.4"
 parking_lot = { version = "0.11" }
 profiling = {version = "1", default-features = false }
 range-alloc = "0.1"
-raw-window-handle = "0.3"
 rend3-types = { version = "^0.0.6", path = "../rend3-types" }
 smallvec = "1"
 smartstring = "0.2"
